refactor(vibes): split style rewrite tests by responsibility
- move style rewrite package tests into focused responsibility folders

- separate output repair, chunking, artifact, model, configuration, input, Chill, and transformer coverage

- share package test doubles from a dedicated support file
